requirements.txt
absl-py                  2.1.0
astunparse               1.6.3
blinker                  1.9.0
cachetools               5.5.0
certifi                  2024.8.30
charset-normalizer       3.4.0
click                    8.1.7
colorama                 0.4.6
Flask                    3.1.0
flatbuffers              24.3.25
gast                     0.6.0
google-api-core          2.23.0
google-auth              2.36.0
google-cloud-bigquery    3.27.0
google-cloud-core        2.4.1
google-crc32c            1.6.0
google-pasta             0.2.0
google-resumable-media   2.7.2
googleapis-common-protos 1.66.0
grpcio                   1.68.1
grpcio-status            1.68.1
h5py                     3.12.1
idna                     3.10
itsdangerous             2.2.0
Jinja2                   3.1.4
keras                    3.7.0
libclang                 18.1.1
Markdown                 3.7
markdown-it-py           3.0.0
MarkupSafe               3.0.2
mdurl                    0.1.2
ml-dtypes                0.4.1
namex                    0.0.8
numpy                    2.0.2
opt_einsum               3.4.0
optree                   0.13.1
packaging                24.2
pip                      24.2
proto-plus               1.25.0
protobuf                 5.29.1
pyasn1                   0.6.1
pyasn1_modules           0.4.1
Pygments                 2.18.0
python-dateutil          2.9.0.post0
python-dotenv            1.0.1
requests                 2.32.3
rich                     13.9.4
rsa                      4.9
setuptools               75.6.0
six                      1.17.0
tensorboard              2.18.0
tensorboard-data-server  0.7.2
tensorflow               2.18.0
tensorflow_intel         2.18.0
termcolor                2.5.0
typing_extensions        4.12.2
urllib3                  2.2.3
Werkzeug                 3.1.3
wheel                    0.45.1
wrapt                    1.17.0